refactor -> GameOver class 구현
