+++ src/main/java/baseball/Controller.java
@@ -5,20 +5,20 @@ import java.util.List;
 import java.util.Scanner;
 import Exception.ProcessException;
 import camp.nextstep.edu.missionutils.Randoms;
-import Exception.CheckException;
 
 public class Controller {
 
     public static void run(Scanner scanner) throws ProcessException {
 
         boolean restartStatus;
+        Game game = new Game(scanner);
+        GameOver gameOver = new GameOver(scanner);
 
         do {
             List<Integer> targetNumber = makeRandomWithoutDuplicate();
             System.out.println(targetNumber);
-            Game game = new Game(scanner);
-            game.play(targetNumber);
-            restartStatus = endingWorking(scanner);
+            game = game.play(targetNumber);
+            restartStatus = gameOver.work();
         } while (restartStatus);
     }
 
@@ -32,17 +32,5 @@ public class Controller {
             }
         }
         return randomValues;
-    }
-
-    private static boolean endingWorking(Scanner scanner) throws IllegalArgumentException{
-
-        System.out.println("3개의 숫자를 모두 맞히셨습니다! 게임 종료\n게임을 새로 시작하려면 1, 종료하려면 2를 입력하세요.");
-        String input = scanner.nextLine();
-        CheckException.restartStatusValid(input);
-        return setRestartStatus(input);
-    }
-
-    private static boolean setRestartStatus(String defineRestartValue) {
-        return defineRestartValue.equals("1");
     }
 }
